Move datadog related code into :datadog module
