+++ app/src/main/AndroidManifest.xml
@@ -322,7 +322,7 @@
                 android:name="com.hedvig.app.feature.tracking.LaunchTrackingInitializer"
                 android:value="androidx.startup" />
             <meta-data
-                android:name="com.hedvig.app.feature.tracking.DatadogInitializer"
+                android:name="com.hedvig.android.datadog.DatadogInitializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
