+++ app/src/main/kotlin/com/hedvig/app/ApplicationModule.kt
@@ -17,9 +17,6 @@ import com.apollographql.apollo3.cache.normalized.normalizedCache
 import com.apollographql.apollo3.interceptor.ApolloInterceptor
 import com.apollographql.apollo3.network.okHttpClient
 import com.apollographql.apollo3.network.ws.SubscriptionWsProtocol
-import com.datadog.android.DatadogEventListener
-import com.datadog.android.DatadogInterceptor
-import com.datadog.android.tracing.TracingInterceptor
 import com.google.firebase.messaging.FirebaseMessaging
 import com.hedvig.android.auth.AuthTokenService
 import com.hedvig.android.auth.interceptor.AuthTokenRefreshingInterceptor
@@ -27,7 +24,9 @@ import com.hedvig.android.auth.interceptor.MigrateTokenInterceptor
 import com.hedvig.android.core.common.di.LogInfoType
 import com.hedvig.android.core.common.di.datastoreFileQualifier
 import com.hedvig.android.core.common.di.isDebugQualifier
+import com.hedvig.android.core.common.di.isProductionQualifier
 import com.hedvig.android.core.common.di.logInfoQualifier
+import com.hedvig.android.datadog.addDatadogConfiguration
 import com.hedvig.android.hanalytics.android.di.appIdQualifier
 import com.hedvig.android.hanalytics.android.di.appVersionCodeQualifier
 import com.hedvig.android.hanalytics.android.di.appVersionNameQualifier
@@ -280,12 +279,6 @@ val applicationModule = module {
   }
 }
 
-private fun OkHttpClient.Builder.addDatadogConfiguration(): OkHttpClient.Builder {
-  return addInterceptor(DatadogInterceptor())
-    .addInterceptor(TracingInterceptor())
-    .eventListenerFactory(DatadogEventListener.Factory())
-}
-
 val apolloClientModule = module {
   single<ApolloClient> {
     val builder: ApolloClient.Builder = get()
@@ -504,6 +497,7 @@ val stringConstantsModule = module {
   single<String>(appVersionCodeQualifier) { BuildConfig.VERSION_CODE.toString() }
   single<String>(appIdQualifier) { BuildConfig.APPLICATION_ID }
   single<Boolean>(isDebugQualifier) { BuildConfig.DEBUG }
+  single<Boolean>(isProductionQualifier) { BuildConfig.BUILD_TYPE == "release" }
 }
 
 val checkoutModule = module {
