+++ core-common/src/main/kotlin/com/hedvig/android/core/common/di/Qualifiers.kt
@@ -8,6 +8,13 @@ import org.koin.core.qualifier.qualifier
 val isDebugQualifier = qualifier("isDebugQualifier")
 
 /**
+ * Provided by the app module, returns whether we are running in release mode.
+ * This is useful as [isDebugQualifier] only returns true for dev environment but false for staging environment and we
+ * do sometimes want to know the difference.
+ */
+val isProductionQualifier = qualifier("buildTypeQualifier")
+
+/**
  * The [java.io.File] to be used from datastore to store the preferences file in.
  * It's provided as a file in order to keep the datastore module free of Android dependencies by skipping fetching the
  * file from context, and instead being fed the file in directly.
