+++ datadog/build.gradle.kts
@@ -2,3 +2,13 @@ plugins {
   id("hedvig.android.ktlint")
   id("hedvig.android.library")
 }
+
+dependencies {
+  implementation(projects.coreCommon)
+
+  implementation(libs.androidx.other.startup)
+  implementation(libs.datadog.sdk)
+  implementation(libs.koin.android)
+  implementation(libs.okhttp.core)
+  implementation(libs.slimber)
+}
