+++ datadog/src/main/kotlin/com/hedvig/android/datadog/DatadogInitializer.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.feature.tracking
+package com.hedvig.android.datadog
 
 import android.content.Context
 import androidx.startup.Initializer
@@ -11,22 +11,25 @@ import com.datadog.android.rum.GlobalRum
 import com.datadog.android.rum.RumMonitor
 import com.datadog.android.rum.tracking.ActivityViewTrackingStrategy
 import com.datadog.android.tracing.AndroidTracer
-import com.hedvig.app.BuildConfig
-import com.hedvig.app.isDebug
-import com.hedvig.app.util.datadog.DatadogLoggingTree
+import com.hedvig.android.core.common.di.isDebugQualifier
+import com.hedvig.android.core.common.di.isProductionQualifier
 import io.opentracing.util.GlobalTracer
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
 import slimber.log.v
 import timber.log.Timber
 
 @Suppress("unused") // Used in /app/src/main/AndroidManifest.xml
-class DatadogInitializer : Initializer<Unit> {
+class DatadogInitializer : Initializer<Unit>, KoinComponent {
+
+  private val isDebug by inject<Boolean>(isDebugQualifier)
+  private val isProduction by inject<Boolean>(isProductionQualifier)
 
   @Suppress("KotlinConstantConditions")
   override fun create(context: Context) {
     val clientToken = "pub185bcba7ed324e83d068b80e25a81359"
     val applicationId = "4d7b8355-396d-406e-b543-30a073050e8f"
 
-    val isProduction = BuildConfig.BUILD_TYPE == "release"
     val environmentName = if (isProduction) "prod" else "dev"
     val configuration = Configuration.Builder(
       logsEnabled = true,
@@ -48,8 +51,8 @@ class DatadogInitializer : Initializer<Unit> {
       rumApplicationId = applicationId,
       serviceName = "android",
     )
-    if (isDebug()) {
-      Datadog.setVerbosity(0)
+    if (isDebug) {
+      Datadog.setVerbosity(android.util.Log.VERBOSE)
     }
     Datadog.initialize(context, credentials, configuration, TrackingConsent.GRANTED)
     val didRegisterGlobalRum = GlobalRum.registerIfAbsent {
